Reduce memory usage for documentation
Do not rely on autoload to source the documentation. We have to
implement the same to check if the doc hash has been initialized, so
just declare the function for `check-declare-file`, then load it through
a regular feature. This allows us to unload any sourced data later with
`unload-feature`.

The savings are miniminal in this case, as we currently keep all strings
except from the initial list container itself.